Solve symbolic link issue for RCP/RAP package (draft)
This first implementation uses a shell script for creating the zip/tar
archives instead of the built-in Tycho product build version that is not
able to create simple symbolic links. The Bash shell script uses
internally the native (Unix/Linux) implementations of zip and tar.

With this addition we loose the ability to run the EPP build on systems
without Bash, zip, tar, gzip.

This is a commit in an early development stage to see if the approach
works on the eclipse.org build servers. More clean-up and the change of
the other packages is required.

See bug 424769: symbolic link missing from "Mac packages"
https://bugs.eclipse.org/bugs/show_bug.cgi?id=424769


Change-Id: I8a4cb60e92492455970ebd0e2d1a99e85e7cf731
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>