Create script that creates a Gerrit verification job settings.xml
When running a Gerrit build verification job, it is the goal to build
only those packages that are affected by a particular change. The shell
script createSettings.sh builds a list with the affected packages based
on the last HEAD commit in the Git repository and creates a special
settings.xml that activates only those packages.

- Changes in packages/ directory will trigger a rebuild of the package 
  with a change.
- Changes in 'packages/*common*' lead to a full rebuild.
- Changes in other areas than in the packages/ directory lead to a full 
  rebuild.

Change-Id: Ia4670058ad5025c34e053521d5db324f7340109e
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>