Bug 462089: The Reporting package refer to outdated TM Terminal features
With TM Terminal 4.0 (Mars), the TM Terminal feature names have changed
and the old deprecated "Terminal" view is no longer available. The
Reporting package should either remove the features or refer to the
replacement for the deprecated "Terminal" view, which is the "Terminals"
view (o.e.tcf.te.terminals.feature).

This change removes the outdated features from the Reporting package.

Change-Id: I32d632e6fab88553d81e8533cdbd868d7f963cb6
Signed-off-by: Uwe Stieber <uwe.stieber@windriver.com>