Bug 462087: The JEE package refer to outdated TM Terminal features
With TM Terminal 4.0 (Mars), the TM Terminal feature names have changed
and the old deprecated "Terminal" view is no longer available. The JEE
package should either remove the features or refer to the replacement
for the deprecated "Terminal" view, which is the "Terminals" view
(o.e.tcf.te.terminals.feature).

This change removes the outdated features from the JavaEE package.

Change-Id: I01cddd3eded278d95f107a2be84b93e16b82c701
Signed-off-by: Uwe Stieber <uwe.stieber@windriver.com>