Disable WindowBuilder dependencies in Java + RCP/RAP package
The artifacts from the WindowBuilder project have been removed in Mars
RC1 from the Simultaneous Release repository. In order to build the
packages the dependencies to WindowBuilder have been disabled. If
WindowBuild will drop out of the release the entries should be finally
removed.

Bug 467800 - WindowBuilder disappeared in Mars RC1
https://bugs.eclipse.org/bugs/show_bug.cgi?id=467800

Change-Id: Ibc26a80bdc9095f9b722b4722730c3c564c779f7
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>