Remove version number from Committer package listing
The Committers package has been listed together with the version number
of the included Eclipse Platform on the main download page. This version
number was useful for those unfamiliar with the names of the
Simultaneous Releases. After some years this doesn't seem the case any
more.

Bug 480642 - Remove version from download title
https://bugs.eclipse.org/bugs/show_bug.cgi?id=480642

Change-Id: I7b52aed0f65cfb7c94348eab42f2adff3e29c4bd
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>