CPP: Move package content definition from feature to product
By using the product definition (instead of the EPP package feature) for
defining the product content it is possible to install some of them as
root level feature that can be updated and/or removed independently from
the package.

The CPP package defines additional (optional) dependencies in the p2.inf
file of the CPP feature; this hasn't been touched as part of this change
and probably needs some work in the future.

Bug 332989 - Allow parts of a package to upgraded or removed
https://bugs.eclipse.org/bugs/show_bug.cgi?id=332989

Change-Id: I81cd464126caae06fb3ba4080bd853035decff4d
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>