DSL: Move package content definition from feature to product
By using the product definition (instead of the EPP package feature) for
defining the product content it is possible to install some of them as
root level feature that can be updated and/or removed independently from
the package.

Since the DSL package already moved some of the dependencies to the
product definition, no need to change anything substantial here, just
doing some minor changes such as adding the EPP common feature to the
product, and adding the platform dependency like in the other packages.

Bug 332989 - Allow parts of a package to upgraded or removed
https://bugs.eclipse.org/bugs/show_bug.cgi?id=332989

Change-Id: I5b28d537a7c810a2be3ab0f5d723e4029dcbdad4
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>