JavaEE: Move package content definition from feature to product
By using the product definition (instead of the EPP package feature) for
defining the product content it is possible to install some of them as
root level feature that can be updated and/or removed independently from
the package.

The JavaEE package defines some additional dependencies to bundles in
its feature; these dependencies haven't been changed as part of this
change.

Bug 332989 - Allow parts of a package to upgraded or removed
https://bugs.eclipse.org/bugs/show_bug.cgi?id=332989

Change-Id: Ib391262bae25956b7f53150f4c87ab6483c47bbe
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>