Modeling: Move package content definition from feature to product
By using the product definition (instead of the EPP package feature) for
defining the product content it is possible to install some of them as
root level feature that can be updated and/or removed independently from
the package.

The features o.e.emf.ecoretools.design o.e.emf.eef.codegen.ecore-feature
are defined including a version number; this should be revised if it is
still necessary and wanted.

Bug 332989 - Allow parts of a package to upgraded or removed
https://bugs.eclipse.org/bugs/show_bug.cgi?id=332989

Change-Id: Iaa2d57ce0f26e0624241952a169ee2ff0a579676
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>