Parallel: Move package content definition from feature to product
By using the product definition (instead of the EPP package feature) for
defining the product content it is possible to install some of them as
root level feature that can be updated and/or removed independently from
the package.

The plug-in includes in the Parallel feature have not been touched;
changing that is not part of this restructuring.

Bug 332989 - Allow parts of a package to upgraded or removed
https://bugs.eclipse.org/bugs/show_bug.cgi?id=332989

Change-Id: I1dd2124b0105d080300d898affa38c902dc8995a
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>