PHP: Move package content definition from feature to product
By using the product definition (instead of the EPP package feature) for
defining the product content it is possible to install some of them as
root level feature that can be updated and/or removed independently from
the package.

One of the main features, org.eclipse.php, is defined including a
version number; this should be revised if it is still necessary and
wanted.

Bug 332989 - Allow parts of a package to upgraded or removed
https://bugs.eclipse.org/bugs/show_bug.cgi?id=332989

Change-Id: I6c1181d0c5a8c23980ff1e8bf5573fd1b7728dc2
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>
Signed-off-by: Kaloyan Raev <kaloyan.r@zend.com>