RCP/RAP: Move package content definition from feature to product
By using the product definition (instead of the EPP package feature) for
defining the product content it is possible to install some of them as
root level feature that can be updated and/or removed independently from
the package.

Only cosmetic changes... the main part of the restructuring was done for
Eclipse Mars - this change removes unnecessary parts and cleans things
up.

Bug 332989 - Allow parts of a package to upgraded or removed
https://bugs.eclipse.org/bugs/show_bug.cgi?id=332989

Change-Id: Ia4919d0c38839b6ef5e17063dda6a89d36144788
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>