Bug 490552 - Move features that are not OS specific to top level
To be able to upgrade features, they need to be at the top level.
Some features used to be installable only on Linux so they had to be
put in p2.inf to be optional but some are now installable eveywhere.
So the features that are not Linux-specific can be moved to the top
level again.

Change-Id: I96351b4be27c618184b88a596416a5ede8998e76
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>