Bug 490341 - Enable StringDeduplication in EPP packages
There are a significant number of duplicated String instances in
Eclipse, primarily from the extension registry and other XML based
files. Using StringDeduplication will allow the underlying character
arrays to be uniqued without requiring an explicit intern operation.
Since StringDeduplication only works when the G1 garbage collector is
used, enable this as well. Both G1 and StringDeduplication were added to
OpenJDK in Java 8, and there is a minium for Java 8 to run the Eclipse
IDE as well as the -Dosgi.requiredJavaVersion specified in these product
files.

All packages have this define this in their common section for Java VM
arguments, *except* the Parallel package that uses OS-specific sections
to define JVM arguments.

Change-Id: I0af43c31b560a2240725b7dce89bba0cd004ef88
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>