Disable Eclipse MPC in Photon M2
The Eclipse Marketplace Client (MPC) is not included in the Photon M2
builds. Because this feature is a mandatory requirement in *all* EPP
packages, we need to temporarily remove this dependency in order to ship
the packages. As soon as it is available, we will activate it again.

Note: This commit removes the dependency only. All other settings in
plugin.xml files etc. remain unchanged (and inactive without the MPC).

Change-Id: I389a077ffb51459a74faee9a7af4a2b0cec10a5f
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>