Remove the deprecated "Modeling Discovery" from the package
This feature from the package is now source of a few problems:
* it requires frequent interactions with the teams publishing the
solutions into the release train and such involvement has been pretty
low in the last few years, leading to a number of un-installable
solutions.
* it introduce a dependency to the Amalgam project, from a plugin point
of view, and from a HTTP request point of view as the catalog is
retrieved from the Amalgam download area.

Considering these issues and the fact that it can be replaced by the use
of the Marketplace client (already enabled in the package) and that the
later has way more features it seems like a good idea to revise this
packaging to remove it.

Change-Id: If026f74262fb56e3275544119fdd6f3ec386d4ae
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>