Bug 535504 - Restore appName to "Eclipse"
The appName is a technical identifier, not part of branding. Changing it
can lead to some errors on some OS and also can create (minor but yet
annoying) behavior change for people upgrading as OS may fail at
identifying the application as it used to.
So reverting it to previous name to avoid useless trouble without
reducing the UX or branding.

Change-Id: I9e740480d5c64e2a497febf9ed95ab298890a9fb
Signed-off-by: Mickael Istria <mistria@redhat.com>
Also-By: Holger Voormann <eclipse@voormann.de>