Remove reporting package
The reporting package has not been published in the last EPP releases on
the download page, and the BIRT project which is the core of this
package does not play an active role in the current Eclipse Simultaneous
Release. Removing the package configuration in order to clean up the
repository and build.

Change-Id: I5640c2e0f6065fac55f0f974271937116e2ae00c
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>