Add bash script to update .forceQualifierUpdate files
The .forceQualifierUpdate files are used to enforce incrementing the
version qualifiers of the various artifacts (bundles, features, but also
the products itself).

The `setGitDate` script helps with this task: It uses UTC noon of the
current day, and writes this into all .forceQualifierUpdate files. If
this script is called with

 . ./setGitDate

it additionally exports two environment variables that force Git to use
the exact same timestamp as Git author and Git commit time. Note that
the build uses the Git commit time for generating the version qualifier.

Change-Id: I735fdca7a78d28d7bb34b28dd4008ac44439a7c9
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>