Remove redundant/obsolete/unnecessary scripts
Note: packages_map.txt and functions.sh was just a way of copying files
from git to the local directory. As the repo is checked out when
promote-a-build is run, we don't need to go back to git again to
get these files. Also, the renaming that happened followed
a very repeatable pattern, so in all we can get rid of two
large files and replace it with a short loop in promote-a-build.

Change-Id: If7acd44d662c7ff9cb7957442cb9be9a430d1cc3